+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 
 * S1 Pattern likelihood in LowEnergyCuts (#21)
 * Max PMT in S1 (v0) LowEnergyCuts (#15)
+* S2AreaFractionTopCut now can have v3 (v2 still default) with tighter AFT selection (#14)
 
 0.4.0 (2017-02-24)
 ------------------